requirements.txt
boto3==1.42.35
botocore==1.42.35
faiss-cpu==1.13.0
jmespath==1.1.0
numpy==2.0.2
packaging==25.0
pypdf==6.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.16.0
six==1.17.0
urllib3==1.26.17
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==11.3.0
streamlit==1.50.0
langsmith
